+++ lib/features/login/data/login_use_case_errors.dart
@@ -1,0 +1,1 @@
+enum LoginUseCaseErrors { connectivity, invalidData, invalidCredential }
