+++ test/login/login_use_case_test.dart
@@ -1,7 +1,9 @@
 import 'dart:convert';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:http/http.dart' as http;
+import 'package:thepos/features/login/data/login_result.dart';
+import 'package:thepos/features/login/data/login_use_case.dart';
+import 'package:thepos/features/login/data/login_use_case_errors.dart';
 
 import '../helpers/mock_client_stub.dart';
 import '../invoice/helpers/shared_test_helper.dart';
@@ -12,74 +14,6 @@ class LoginUseCaseSUT {
   final MockClientStub client;
   final LoginUseCase loginUseCase;
 }
-
-class LoginResult {
-  LoginResult({
-    required this.token,
-    required this.user,
-    required this.expire,
-    required this.displayName,
-  });
-
-  factory LoginResult.fromJson(Map<String, dynamic> json) => LoginResult(
-        token: json['token'],
-        user: json['user'],
-        expire: json['expire'],
-        displayName: json['display_name'],
-      );
-
-  String token;
-  String user;
-  int expire;
-  String displayName;
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = <String, dynamic>{};
-    data['token'] = token;
-    data['user'] = user;
-    data['expire'] = expire;
-    data['display_name'] = displayName;
-    return data;
-  }
-}
-
-class LoginUseCase {
-  LoginUseCase(this._client, this._url);
-
-  final http.Client _client;
-  final Uri _url;
-
-  Future<LoginResult> login(String username, String password) async {
-    try {
-      final http.Response response = await _client.post(
-        _url,
-        body: json.encode(<String, String>{
-          'username': username,
-          'password': password,
-        }),
-      );
-      if (response.statusCode == 200) {
-        return _tryParse(response.body);
-      } else if (response.statusCode == 401) {
-        return Future<LoginResult>.error(LoginUseCaseErrors.invalidCredential);
-      } else {
-        return Future<LoginResult>.error(LoginUseCaseErrors.invalidData);
-      }
-    } catch (error) {
-      return Future<LoginResult>.error(LoginUseCaseErrors.connectivity);
-    }
-  }
-
-  Future<LoginResult> _tryParse(String body) {
-    try {
-      return Future<LoginResult>.value(LoginResult.fromJson(jsonDecode(body)));
-    } catch (error) {
-      return Future<LoginResult>.error(LoginUseCaseErrors.invalidData);
-    }
-  }
-}
-
-enum LoginUseCaseErrors { connectivity, invalidData, invalidCredential }
 
 void main() {
   LoginUseCaseSUT _makeSUT() {
